PR feedback adjustments/fixes
