+++ Kwartracker/Custom Components/View/View+Extensions.swift
@@ -9,6 +9,6 @@ import SwiftUI
 
 extension View {
     func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape( RoundedCorner(radius: radius, corners: corners) )
+        clipShape(RoundedCorner(radius: radius, corners: corners))
     }
 }
